mon/Elector: make paxos_size() an int
We want to get rid of the warning:

‘int’ and ‘unsigned int’ [-Wsign-compare]
if (rank_removed < paxos_size()) {
~~~~~~~~~~~~^~~~~~~~~~~~~

Also, according to google style guide, we should
"try to avoid unsigned types and not use an
unsigned type merely to assert that a variable
is non-negative".

https://google.github.io/styleguide/cppguide.html#Integer_Types

Fixes: https://tracker.ceph.com/issues/56392

+++ src/mon/Elector.cc
@@ -130,9 +130,9 @@ bool Elector::ever_participated() const
   return mon->has_ever_joined;
 }
 
-unsigned Elector::paxos_size() const
+int Elector::paxos_size() const
 {
-  return (unsigned)mon->monmap->size();
+  return mon->monmap->size();
 }
 
 void Elector::shutdown()
